requirements.txt
openai
seatable-api
python-dotenv
playwright
tqdm
requests
beautifulsoup4
